chore: Separando código de de registro e login
