+++ src/pages/Login/index.tsx
@@ -1,23 +1,15 @@
-import React from 'react';
+import React, { useState } from 'react';
+
+import { SignIn } from './SignIn';
+import { SignUp } from './SignUp';
 
 import { Container } from './style';
 
-import FormControl from '@mui/material/FormControl';
-import Visibility from '@mui/icons-material/Visibility';
-import VisibilityOff from '@mui/icons-material/VisibilityOff';
-import InputAdornment from '@mui/material/InputAdornment';
-import OutlinedInput from '@mui/material/OutlinedInput';
-import InputLabel from '@mui/material/InputLabel';
-import IconButton from '@mui/material/IconButton';
-import MailOutlineIcon from '@mui/icons-material/MailOutline';
-import LockOpenIcon from '@mui/icons-material/LockOpen';
-import Button from '@mui/material/Button';
+import Box from '@mui/material/Box';
 import PropTypes from 'prop-types';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Typography from '@mui/material/Typography';
-import Box from '@mui/material/Box';
-import PersonOutlinedIcon from '@mui/icons-material/PersonOutlined';
 
 interface TabPanelProps {
     className?: string;
@@ -62,13 +54,9 @@ function a11yProps(index: number) {
 
 export function Login() {
 
-    const [value, setValue] = React.useState(0);
-    const [showPassword, setShowPassword] = React.useState(false);
-    
+    const [value, setValue] = useState(0);
     const handleChange = (_event: React.SyntheticEvent, newValue: number) => {setValue(newValue);};
-    const handleClickShowPassword = () => setShowPassword((show) => !show);
-    const handleMouseDownPassword = (event: React.MouseEvent<HTMLButtonElement>) => {event.preventDefault();};
-  
+
     return (
         <Container className='flex'>
             <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
@@ -78,131 +66,10 @@ export function Login() {
                 </Tabs>
             </Box>
             <CustomTabPanel value={value} index={0}>
-                <form className='flex'>
-                    <FormControl>
-                        <InputLabel htmlFor="outlined-adornment-email">E-mail</InputLabel>
-                        <OutlinedInput
-                            id="outlined-adornment-email"
-                            placeholder='Digite seu E-mail'
-                            type='text'
-                            startAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton edge="start">
-                                        <MailOutlineIcon />
-                                    </IconButton>
-                                </InputAdornment>
-                            }
-                            endAdornment={
-                                <InputAdornment position="end">
-                                    <p>@acesse.net.br</p>
-                                </InputAdornment>
-                            }
-                            label="E-mail"
-                        />
-                    </FormControl>
-                    <FormControl>
-                        <InputLabel htmlFor="outlined-adornment-password">Password</InputLabel>
-                        <OutlinedInput
-                            id="outlined-adornment-password"
-                            placeholder='Digite sua senha'
-                            type={showPassword ? 'text' : 'password'}
-                            startAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton edge="start">
-                                        <LockOpenIcon />
-                                    </IconButton>
-                                </InputAdornment>
-                            }
-                            endAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton
-                                        aria-label="toggle password visibility"
-                                        onClick={handleClickShowPassword}
-                                        onMouseDown={handleMouseDownPassword}
-                                        edge="end"
-                                    >
-                                        {showPassword ? <VisibilityOff /> : <Visibility />}
-                                    </IconButton>
-                                </InputAdornment>
-                            }
-                            label="Password"
-                        />
-                    </FormControl>
-                    <Button variant="contained" size="large">
-                        Entrar
-                    </Button>
-                </form>
+                <SignIn />
             </CustomTabPanel>
             <CustomTabPanel value={value} index={1}>
-                <form className='flex'>
-                    <FormControl>
-                        <InputLabel htmlFor="outlined-adornment-user">Usuário</InputLabel>
-                        <OutlinedInput
-                            id="outlined-adornment-user"
-                            placeholder='Digite seu nome'
-                            startAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton edge="start">
-                                        <PersonOutlinedIcon />
-                                    </IconButton>
-                                </InputAdornment>
-                            }   
-                            label="Usuário"
-                        />
-                    </FormControl>
-                    <FormControl>
-                        <InputLabel htmlFor="outlined-adornment-password">E-mail</InputLabel>
-                        <OutlinedInput
-                            id="outlined-adornment-password"
-                            placeholder='Digite seu E-mail'
-                            type='text'
-                            startAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton edge="start">
-                                        <MailOutlineIcon />
-                                    </IconButton>
-                                </InputAdornment>
-                            }
-                            endAdornment={
-                                <InputAdornment position="end">
-                                    <div>@acesse.net.br</div>
-                                </InputAdornment>
-                            }
-                            label="E-mail"
-                        />
-                    </FormControl>
-                    <FormControl>
-                        <InputLabel htmlFor="outlined-adornment-password">Password</InputLabel>
-                        <OutlinedInput
-                            id="outlined-adornment-password"
-                            placeholder='Digite sua senha'
-                            type={showPassword ? 'text' : 'password'}
-                            startAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton edge="start">
-                                        <LockOpenIcon />
-                                    </IconButton>
-                                </InputAdornment>
-                            }
-                            endAdornment={
-                                <InputAdornment position="end">
-                                    <IconButton
-                                        aria-label="toggle password visibility"
-                                        onClick={handleClickShowPassword}
-                                        onMouseDown={handleMouseDownPassword}
-                                        edge="end"
-                                    >
-                                        {showPassword ? <VisibilityOff /> : <Visibility />}
-                                    </IconButton>
-                                </InputAdornment>
-                            }
-                            label="Password"
-                        />
-                    </FormControl>
-                    <Button variant="contained" size="large">
-                        Registrar
-                    </Button>
-                </form>
+                <SignUp />
             </CustomTabPanel>
         </Container>
     )
